[#178848443] Rename "action" tab to "prioritize-basins" tab
