+++ src/components/pages/map/component.jsx
@@ -9,7 +9,7 @@ import { toastr } from 'react-redux-toastr';
 // components
 import BaselineTab from 'components/pages/map/baseline-tab';
 import FutureTab from 'components/pages/map/future-tab';
-import ActionTab from 'components/pages/map/action-tab';
+import PrioritizeBasinsTab from 'components/pages/map/prioritize-basins-tab';
 import MapComponent from 'components/map';
 import CustomAccordion from 'components/ui/custom-accordion';
 import Analyzer from 'components/analyzer';
@@ -132,7 +132,7 @@ class MapPage extends PureComponent {
           <div className="l-mapview-content">
             {scope === 'baseline' && (<BaselineTab />)}
             {scope === 'future' && (<FutureTab />)}
-            {scope === 'action' && (<ActionTab />)}
+            {scope === 'prioritize-basins' && (<PrioritizeBasinsTab />)}
           </div>
           <CustomAccordion
             className="l-analyzer-accordion"
